moved calculations to aggregation table
